# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CyberArcade CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contacts/{id}/campaigns:
    get:
      operationId: getContactCampaigns
      tags: [contacts]
      summary: Get campaigns a contact is part of
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CampaignLog"
...
  /campaigns:
    get:
      operationId: getCampaigns
      tags: [campaigns]
      summary: List campaigns
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignInput"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /campaigns/{id}/launch:
    post:
      operationId: launchCampaign
      tags: [campaigns]
      summary: Launch a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Launched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/pause:
    post:
      operationId: pauseCampaign
      tags: [campaigns]
      summary: Pause a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/duplicate:
    post:
      operationId: duplicateCampaign
      tags: [campaigns]
      summary: Duplicate a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Duplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/logs:
    get:
      operationId: getCampaignLogs
      tags: [campaigns]
      summary: Get campaign logs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CampaignLog"
  /campaigns/{id}/timeline:
    get:
      operationId: getCampaignTimeline
      tags: [campaigns]
      summary: Get campaign engagement timeline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Timeline data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelinePoint"
...
components:
  schemas:
...
    DeleteResult:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    Campaign:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        goal:
          type: string
          enum: [enroll_trial, new_acquisition, re_engage, batch_announce, awareness]
        status:
          type: string
          enum: [draft, scheduled, active, paused, completed]
        channels:
          type: array
          items:
            type: string
        segmentId:
          type: ["string", "null"]
        segmentName:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        messageBody:
          type: ["string", "null"]
        ctaText:
          type: ["string", "null"]
        ctaUrl:
          type: ["string", "null"]
        sentCount:
          type: integer
        openCount:
          type: integer
        clickCount:
          type: integer
        replyCount:
          type: integer
        conversionCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - goal
        - status
        - channels
        - sentCount
        - openCount
        - clickCount
        - replyCount
        - conversionCount
        - createdAt
    CreateCampaignInput:
      type: object
      properties:
        name:
          type: string
        goal:
          type: string
          enum: [enroll_trial, new_acquisition, re_engage, batch_announce, awareness]
        status:
          type: string
          enum: [draft, scheduled, active, paused, completed]
        channels:
          type: array
          items:
            type: string
        segmentId:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        messageBody:
          type: ["string", "null"]
        ctaText:
          type: ["string", "null"]
        ctaUrl:
          type: ["string", "null"]
      required:
        - name
        - goal
        - channels
    UpdateCampaignInput:
      type: object
      properties:
        name:
          type: string
        goal:
          type: string
          enum: [enroll_trial, new_acquisition, re_engage, batch_announce, awareness]
        status:
          type: string
          enum: [draft, scheduled, active, paused, completed]
        channels:
          type: array
          items:
            type: string
        segmentId:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
        subject:
          type: ["string", "null"]
        messageBody:
          type: ["string", "null"]
        ctaText:
          type: ["string", "null"]
        ctaUrl:
          type: ["string", "null"]
        sentCount:
          type: integer
        openCount:
          type: integer
        clickCount:
          type: integer
        replyCount:
          type: integer
        conversionCount:
          type: integer
    CampaignLog:
      type: object
      properties:
        id:
          type: string
        campaignId:
          type: string
        contactId:
          type: string
        contactName:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        channel:
          type: string
        status:
          type: string
          enum: [sent, delivered, opened, clicked, replied, converted, failed]
        sentAt:
          type: ["string", "null"]
        openedAt:
          type: ["string", "null"]
        clickedAt:
          type: ["string", "null"]
        convertedAt:
          type: ["string", "null"]
      required:
        - id
        - campaignId
        - contactId
        - channel
        - status
    TimelinePoint:
      type: object
      properties:
        date:
          type: string
        opens:
          type: integer
        clicks:
          type: integer
        conversions:
          type: integer
      required:
        - date
        - opens
        - clicks
        - conversions
...