# lib/api-zod/src/generated/types/timelinePoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CyberArcade CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface TimelinePoint {
  date: string;
  opens: number;
  clicks: number;
  conversions: number;
}